Changes in token handling; test spliting
