+++ lib/json2sql.js
@@ -3,31 +3,33 @@ function json2sql() {}
 
 const parseFunction = (nodeFun) => {
     const args = [];
-    for (let i = 0, length = nodeFun.arguments.length; i < length; i++) {
-        const node = nodeFun.arguments[i];
-        switch (node.type) {
+    if (nodeFun.arguments) {
+        for (let i = 0, length = nodeFun.arguments.length; i < length; i++) {
+            const node = nodeFun.arguments[i];
+            switch (node.type) {
 
-            case 'literal':
-                args.push(node.value);
-                break;
-            case 'string':
-                args.push(`'${node.value}'`);
-                break;
-            case 'wildcard':
-                args.push(`*`);
-                break;
-            case 'number':
-                args.push(node.value);
-                break;
-            case 'math':
-                args.push(parseMath(node));
-                break;
-            case 'function':
-                args.push(parseFunction(node));
-                break;
-            default:
-                break;
+                case 'literal':
+                    args.push(`"${node.value}"`);
+                    break;
+                case 'string':
+                    args.push(`'${node.value}'`);
+                    break;
+                case 'wildcard':
+                    args.push(`*`);
+                    break;
+                case 'number':
+                    args.push(node.value);
+                    break;
+                case 'math':
+                    args.push(parseMath(node));
+                    break;
+                case 'function':
+                    args.push(parseFunction(node));
+                    break;
+                default:
+                    break;
 
+            }
         }
     }
     return `${nodeFun.value}(${args.join(',')})${nodeFun.alias ? ` AS ${nodeFun.alias}` : ''}`;
@@ -40,7 +42,7 @@ const parseMath = (nodeFun) => {
         switch (node.type) {
 
             case 'literal':
-                args.push(node.value);
+                args.push(`"${node.value}"`);
                 break;
             case 'string':
                 args.push(`'${node.value}'`);
@@ -70,10 +72,10 @@ const parseSelect = (select) => {
                     responses.push(`*`);
                     break;
                 case 'literal':
-                    responses.push(`${node.value}${node.alias ? ` AS ${node.alias}` : ''}`);
+                    responses.push(`"${node.value}"${node.alias ? ` AS ${node.alias}` : ''}`);
                     break;
                 case 'string':
-                    responses.push(`${node.value}${node.alias ? ` AS ${node.alias}` : ''}`);
+                    responses.push(`'${node.value}'${node.alias ? ` AS ${node.alias}` : ''}`);
                     break;
                 case 'function':
                     responses.push(parseFunction(node));
@@ -86,6 +88,9 @@ const parseSelect = (select) => {
                     break;
                 case 'distinct':
                     responses.push(`DISTINCT ${parseSelect(node.arguments)}`);
+                    break;
+                case 'boolean':
+                    responses.push(`${node.value}${node.alias ? ` AS ${node.alias}` : ''}`);
                     break;
                 default:
                     break;
@@ -104,7 +109,7 @@ const parseNodeWhere = (node) => {
         case 'number':
             return node.value;
         case 'string':
-            return `${node.value}`;
+            return `'${node.value}'`;
         case 'operator':
             return `${parseNodeWhere(node.left)} ${node.value} ${parseNodeWhere(node.right)}`;
         case 'conditional':
@@ -152,10 +157,15 @@ const parseOrderBy = (orderBy) => {
     if (orderBy) {
         const responses = [];
         for (let i = 0, length = orderBy.length; i < length; i++) {
-            if (orderBy[i].type === 'function') {
-                responses.push(`${parseFunction(orderBy[i])}${orderBy[i].direction ? ` ${orderBy[i].direction}` : ''}`)
-            } else {
-                responses.push(`${orderBy[i].value}${orderBy[i].direction ? ` ${orderBy[i].direction}` : ''}`);
+            switch (orderBy[i].type) {
+                case 'function':
+                    responses.push(`${parseFunction(orderBy[i])}${orderBy[i].direction ? ` ${orderBy[i].direction}` : ''}`);
+                    break;
+                case 'literal':
+                    responses.push(`"${orderBy[i].value}"${orderBy[i].direction ? ` ${orderBy[i].direction}` : ''}`);
+                    break;
+                default:
+                    responses.push(`${orderBy[i].value}${orderBy[i].direction ? ` ${orderBy[i].direction}` : ''}`);
             }
         }
         return `ORDER BY ${responses.join(', ')}`;
