+++ lib/sql2json.js
@@ -22,7 +22,11 @@ function sql2json(sql, experimental = false) {
     }
     this.sql = sql;
     this.experimental = experimental;
-    this.tokens = lexer.tokenize(this.sql);
+    try {
+        this.tokens = lexer.tokenize(this.sql);
+    } catch (err) {
+        throw new Error('SQL tokenizer detected an error: ' + err.toString());
+    }
     this.index = 0;
     this.stack = [];
     this.parsed = {};
@@ -86,14 +90,9 @@ sql2json.prototype.parseFunction = function (functionName, isInSelect) {
                     arguments: stack
                 };
             case 'STRING':
-                stack.push({
-                    value: `'${token[1]}'`,
-                    type: 'string'
-                });
-                break;
             case 'DBLSTRING':
                 stack.push({
-                    value: `"${token[1]}"`,
+                    value: `${token[1]}`,
                     type: 'string'
                 });
                 break;
@@ -187,7 +186,7 @@ sql2json.prototype.parseSelect = function () {
         switch (obtainType(token, this.lookAhead())) {
             case 'DBLSTRING':
                 this.parsed.select.push({
-                    value: `"${token[1]}"`,
+                    value: `${token[1]}`,
                     type: 'string'
                 });
                 break;
@@ -229,7 +228,7 @@ sql2json.prototype.parseSelect = function () {
                     containAs = false;
                 } else {
                     this.parsed.select.push({
-                        value: `'${token[1]}'`,
+                        value: `${token[1]}`,
                         alias: null,
                         type: 'string'
                     });
@@ -237,7 +236,11 @@ sql2json.prototype.parseSelect = function () {
                 break;
             case 'MATH_MULTI':
             case 'MATH':
-                this.parsed.select.push(this.parseMath(token[1], this.parsed.select.pop()));
+                if (this.parsed.select.length > 0) {
+                    this.parsed.select.push(this.parseMath(token[1], this.parsed.select.pop()));
+                } else {
+                    throw new Error('MATH operation found with empty stack - is your query correctly formed?')
+                }
                 break;
             case 'AS':
                 containAs = true;
@@ -315,9 +318,10 @@ sql2json.prototype.parseIn = function () {
             case 'LEFT_PAREN':
             case 'SEPARATOR':
                 break;
+            case 'DBLSTRING':
             case 'STRING':
                 stack.push({
-                    value: `'${token[1]}'`,
+                    value: `${token[1]}`,
                     type: 'string'
                 });
                 break;
@@ -483,14 +487,9 @@ sql2json.prototype.parseWhere = function () {
                 minus = false;
                 break;
             case 'DBLSTRING':
-                stack.push({
-                    value: `"${token[1]}"`,
-                    type: 'string'
-                });
-                break;
             case 'STRING':
                 stack.push({
-                    value: `'${token[1]}'`,
+                    value: `${token[1]}`,
                     type: 'string'
                 });
                 break;
@@ -632,9 +631,10 @@ sql2json.prototype.parseOrder = function () {
                 break;
             case 'SEPARATOR':
                 break;
+            case 'STRING':
             case 'DBLSTRING':
                 stack.push({
-                    value: `"${token[1]}"`,
+                    value: `${token[1]}`,
                     alias: null,
                     type: 'string',
                     direction: null
@@ -827,7 +827,8 @@ sql2json.prototype.parseGroupExperimental = function () {
 sql2json.prototype.parse = function () {
     while (this.hasNext()) {
         const token = this.next();
-        switch (obtainType(token, this.lookAhead())) {
+        const tokenType = obtainType(token, this.lookAhead());
+        switch (tokenType) {
 
             case 'SELECT':
                 if (this.experimental) {
@@ -872,6 +873,7 @@ sql2json.prototype.parse = function () {
             case 'EOF':
                 return;
             default:
+                throw new Error('Unsupported query element detected: ' + token[1]);
                 break;
 
         }
