+++ test/json2sql/group-by.js
@@ -1,0 +1,123 @@
+const assert = require('assert');
+const Json2sql = require('../../index').json2sql;
+require('should');
+
+
+describe('JSON to SQL - GroupBy', () => {
+    it('Group by one field', () => {
+        const data = {
+            select: [{
+                value: '*',
+                alias: null,
+                type: 'wildcard'
+            }],
+            from: 'tablename',
+            group: [{
+                type: 'literal',
+                value: 'name'
+            }]
+        };
+
+        const response = 'SELECT * FROM tablename GROUP BY name';
+        Json2sql.toSQL(data).should.deepEqual(response);
+    });
+
+    it('Group by number of column name', () => {
+        const data = {
+            select: [{
+                value: '*',
+                alias: null,
+                type: 'wildcard'
+            }],
+            from: 'tablename',
+            group: [{
+                type: 'number',
+                value: 123
+            }]
+        };
+
+        const response = 'SELECT * FROM tablename GROUP BY 123';
+        Json2sql.toSQL(data).should.deepEqual(response);
+    });
+
+    it('Group by several fields', () => {
+        const data = {
+            select: [{
+                value: '*',
+                alias: null,
+                type: 'wildcard'
+            }],
+            from: 'tablename',
+            group: [{
+                type: 'literal',
+                value: 'name'
+            }, {
+                type: 'literal',
+                value: 'surname'
+            }]
+        };
+
+        const response = 'SELECT * FROM tablename GROUP BY name, surname';
+        Json2sql.toSQL(data).should.deepEqual(response);
+    });
+
+    it('Group with where', () => {
+        const data = {
+            select: [{
+                value: '*',
+                alias: null,
+                type: 'wildcard'
+            }],
+            from: 'tablename',
+            group: [{
+                type: 'literal',
+                value: 'name'
+            }, {
+                type: 'literal',
+                value: 'surname'
+            }],
+            where: {
+                type: 'between',
+                value: 'data',
+                arguments: [{
+                    value: 1,
+                    type: 'number'
+                }, {
+                    value: 3,
+                    type: 'number'
+                }]
+            }
+        };
+
+        const response = 'SELECT * FROM tablename WHERE data BETWEEN 1 AND 3 GROUP BY name, surname';
+        Json2sql.toSQL(data).should.deepEqual(response);
+    });
+
+    it('Group with function', () => {
+        const data = {
+            select: [{
+                value: '*',
+                alias: null,
+                type: 'wildcard'
+            }],
+            from: 'tablename',
+            group: [{
+                type: 'function',
+                value: 'ST_GeoHash',
+                alias: null,
+                arguments: [{
+                    type: 'literal',
+                    value: 'the_geom_point'
+                }, {
+                    type: 'number',
+                    value: 8
+                }]
+            }],
+
+        };
+
+        const response = 'SELECT * FROM tablename GROUP BY ST_GeoHash("the_geom_point",8)';
+        Json2sql.toSQL(data).should.deepEqual(response);
+    });
+});
+
